--- src/comments/comment.entity.ts
@@ -1,55 +1,0 @@
-import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity()
-export class Comment {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  username: string;
-  
-  @Column()
-  date: string;
-
-  @Column({type : "int", width: 5})
-  star: number;
-
-  @Column()
-  like: number;
-
-  @Column()
-  title: string;
-
-  @Column()
-  review: string;
-  
-}
-
-// import { Document } from "mongoose";
-// import { Prop, Schema, SchemaFactory} from '@nestjs/mongoose'
-
-// export type CommentDocument = Comment & Document;
-
-
-// @Schema()
-// export class Comment {
-//     @Prop()
-//     user: string;
-
-//     @Prop()
-//     date: { type: Date };
-
-//     @Prop()
-//     star: {type : Number, min:0, max:5};
-
-//     @Prop()
-//     like: number;
-
-//     @Prop()
-//     title: string;
-
-//     @Prop()
-//     review: string;
-// }
-
-// export const CommentSchema = SchemaFactory.createForClass(Comment);
