--- src/comments/comment.schema.ts
@@ -1,28 +1,0 @@
-// import { Document } from "mongoose";
-// import { Prop, Schema, SchemaFactory} from '@nestjs/mongoose'
-
-// export type CommentDocument = Comment & Document;
-
-
-// @Schema()
-// export class Comment {
-//     @Prop()
-//     user: string;
-
-//     @Prop()
-//     date: { type: Date };
-
-//     @Prop()
-//     star: {type : Number, min:0, max:5};
-
-//     @Prop()
-//     like: number;
-
-//     @Prop()
-//     title: string;
-
-//     @Prop()
-//     review: string;
-// }
-
-// export const CommentSchema = SchemaFactory.createForClass(Comment);
