+++ src/review/review.entity.ts
@@ -1,0 +1,27 @@
+import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity()
+export class Comment {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  username: string;
+  
+  @Column()
+  date: string;
+
+  @Column({type : "int", width: 5})
+  star: number;
+
+  @Column()
+  like: number;
+
+  @Column()
+  title: string;
+
+  @Column()
+  review: string;
+  
+}
+// export const CommentSchema = SchemaFactory.createForClass(Comment);
